Organized cypress integration page tests into folders to verify if 'MaxListenersExceededWarning: Possible EventEmitter memory leak detected' is resolved
