+++ test/cypress/integration/pages/consumer-tools/tdp-activity-search.js
@@ -1,4 +1,4 @@
-import { ActivitySearch } from '../../pages/practitioner-resources/activity-search';
+import { ActivitySearch } from '../../pages/consumer-tools/activity-search';
 
 const search = new ActivitySearch();
