+++ test/cypress/integration/pages/data-research/credit-card-agreements-search.js
@@ -1,4 +1,4 @@
-import { CreditCardAgreementSearch } from '../../pages/credit-cards/agreements-search';
+import { CreditCardAgreementSearch } from '../../../pages/data-research/credit-card-agreements-search';
 
 const searchPage = new CreditCardAgreementSearch();
