+++ test/cypress/integration/pages/data-research/mortgage-performance-trends.js
@@ -1,4 +1,4 @@
-import { MortgagePerformanceTrends } from '../../pages/data-research/mortgage-performance-trends';
+import { MortgagePerformanceTrends } from '../../../pages/data-research/mortgage-performance-trends';
 
 const trends = new MortgagePerformanceTrends();
