+++ test/cypress/integration/pages/paying-for-college/choose-a-student-loan.js
@@ -1,4 +1,4 @@
-import { PayingForCollege } from '../../components/paying-for-college.js';
+import { PayingForCollege } from '../../../components/paying-for-college.js';
 
 const page = new PayingForCollege();
