+++ test/cypress/integration/pages/paying-for-college/manage-your-college-money.js
@@ -1,4 +1,4 @@
-import { PayingForCollege } from '../../components/paying-for-college.js';
+import { PayingForCollege } from '../../../components/paying-for-college.js';
 
 const page = new PayingForCollege();
