+++ test/cypress/integration/pages/rules-policy/regulations-search.js
@@ -1,4 +1,4 @@
-import { RegulationsSearch } from '../../pages/policy-compliance/regulations-search';
+import { RegulationsSearch } from '../../../pages/rules-policy/regulations-search';
 
 const regulationsSearch = new RegulationsSearch();
